imported DB functions and respond to GET with appropriate data
